# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/item-modifier/items/create": {
      "post": {
        "operationId": "ItemModifierController_createItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
...
    "/item-modifier/items/update/{id}": {
      "patch": {
        "operationId": "ItemModifierController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/items/delete/{id}": {
      "delete": {
        "operationId": "ItemModifierController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/item-groups/create": {
      "post": {
        "operationId": "ItemModifierController_createItemGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
...
    "/item-modifier/item-groups/update/{id}": {
      "patch": {
        "operationId": "ItemModifierController_updateItemGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/item-groups/delete/{id}": {
      "delete": {
        "operationId": "ItemModifierController_removeItemGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/modifiers/create": {
      "post": {
        "operationId": "ItemModifierController_createModifier",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModifierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
...
    "/item-modifier/modifiers/update/{id}": {
      "patch": {
        "operationId": "ItemModifierController_updateModifier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModifierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/modifiers/delete/{id}": {
      "delete": {
        "operationId": "ItemModifierController_removeModifier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/modifier-groups/create": {
      "post": {
        "operationId": "ItemModifierController_createModifierGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModifierGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
...
    "/item-modifier/modifier-groups/update/{id}": {
      "patch": {
        "operationId": "ItemModifierController_updateModifierGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModifierGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
    "/item-modifier/modifier-groups/delete/{id}": {
      "delete": {
        "operationId": "ItemModifierController_removeModifierGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ItemModifier"
        ]
      }
    },
...
  },
  "info": {
    "title": "OsonPOS API",
    "description": "The OsonPOS API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {}
      },
      "CreateItemGroupDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateItemGroupDto": {
        "type": "object",
        "properties": {}
      },
      "CreateModifierDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateModifierDto": {
        "type": "object",
        "properties": {}
      },
      "CreateModifierGroupDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateModifierGroupDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}